% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-apply_bcftools_norm.R
\name{apply_bcftools_norm}
\alias{apply_bcftools_norm}
\title{Normalize VCF file using bcftools norm
This function normalizes a VCF file by calling bcftools norm on a temp VCF.
The normalized output is then read into a dataframe, extracting the columns:
chr, pos, ref, and alt.}
\usage{
apply_bcftools_norm(chr, pos, ref, alt, fasta, tmp_folder)
}
\arguments{
\item{chr}{A string representing the chromosome.}

\item{pos}{An integer representing the position.}

\item{ref}{A string representing the reference allele.}

\item{alt}{A string representing the alternative allele.}

\item{fasta}{Character vector for the folder temporary path.}

\item{tmp_folder}{Character vector for the folder temporary path.}
}
\value{
A dataframe containing normalized variant information with bcftools.
}
\description{
Normalize VCF file using bcftools norm
This function normalizes a VCF file by calling bcftools norm on a temp VCF.
The normalized output is then read into a dataframe, extracting the columns:
chr, pos, ref, and alt.
}
\keyword{internal}
