% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bitvalues_from_bam_flag}
\alias{bitvalues_from_bam_flag}
\title{Get bit values from bit names of BAM flag.}
\usage{
bitvalues_from_bam_flag(flag, bitnames)
}
\arguments{
\item{flag}{an integer value}

\item{bitnames}{an integer vector}
}
\value{
a matrix with column names matching bitnames and binary values representing bit values
}
\description{
Get bit values from bit names of BAM flag.
}
\keyword{internal}
