% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper.R
\name{build_dir_structure}
\alias{build_dir_structure}
\title{Helper function to build a directory structure for testing
it use files in inst/extdata}
\usage{
build_dir_structure(root = getwd())
}
\arguments{
\item{root}{the root directory to build the structure dirs}
}
\description{
Helper function to build a directory structure for testing
it use files in inst/extdata
}
