% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_bcftools_norm.R
\name{check_bcftools_is_installed}
\alias{check_bcftools_is_installed}
\title{Check that bcftools system dependency is available.}
\usage{
check_bcftools_is_installed()
}
\description{
Check that bcftools system dependency is available.
}
\keyword{internal}
