% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-sanity-checks-input.R
\name{check_input}
\alias{check_input}
\title{Check input files for fRagmentomics function}
\usage{
check_input(
  mut,
  bam,
  fasta,
  sample_id,
  neg_offset_mate_search,
  pos_offset_mate_search,
  one_based,
  flag_keep,
  flag_remove,
  report_tlen,
  report_softclip,
  report_5p_3p_bases_fragment,
  tmp_folder,
  n_cores
)
}
\value{
None. The function stops execution if error.
}
\description{
Check input files for fRagmentomics function
}
\keyword{internal}
