% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{Check and validate all input parameters for the analysis pipeline}
\usage{
check_parameters(
  mut,
  bam,
  fasta,
  sample_id,
  neg_offset_mate_search,
  pos_offset_mate_search,
  one_based,
  flag_bam_list,
  report_tlen,
  report_softclip,
  report_5p_3p_bases_fragment,
  remove_softclip,
  retain_fail_qc,
  tmp_folder,
  output_folder,
  n_cores
)
}
\arguments{
\item{mut}{Path to a .vcf or .tsv file or string representation chr:pos:ref:alt of a mutation.}

\item{bam}{Path to a BAM file.}

\item{fasta}{Path to the FASTA file for the reference sequence used for generating the BAM file.}

\item{sample_id}{Sample identifier.}

\item{neg_offset_mate_search}{Integer. Use in read_bam.
Represents the number of nucleotides to extend upstream (negative direction)
from the position of interest when querying the BAM file with Rsamtools.
his extension ensures that paired reads are retrieved, even if only one mate
overlaps the queried position.}

\item{pos_offset_mate_search}{Integer. Use in read_bam.}

\item{one_based}{Boolean. TRUE if fasta is in one based. False if in 0 based.}

\item{flag_bam_list}{A named list of logicals for filtering reads based on their SAM flag
NA = Filter is ignored, TRUE = The read MUST have this flag, FALSE = The read MUST NOT have this flag.}

\item{report_tlen}{Boolean. Whether to include the TLEN (template length)
information in the output.}

\item{report_softclip}{Boolean. Whether to include the number of soft-clipped
bases at the fragment extremities in the output.}

\item{report_5p_3p_bases_fragment}{Integer. Whether to include N fragment
extremity bases in the output.}

\item{remove_softclip}{Boolean. If set to TRUE, trim soft-clipped bases from the 5' end of Read 5p and from the 3' end of Read 3p.}

\item{retain_fail_qc}{Boolean. If set to TRUE, retain fragments that failed the various quality checks in the output.}

\item{tmp_folder}{Character vector for the temporary folder path.}

\item{n_cores}{Number of cores for parallel computation.}
}
\value{
None. The function stops execution if error.
}
\description{
This function serves as a centralized gatekeeper, validating all user-provided parameters before the main
analysis begins. It calls a series of specialized  helper functions to check each parameter for correctness (e.g., file existence,
data type, valid values) and stops execution with an informative error message if any check fails.
}
\keyword{internal}
