% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_seq_rep}
\alias{check_seq_rep}
\title{Check sequence repetition}
\usage{
check_seq_rep(indel_rep, is_length_ok, diff_mutation_bam, deletion_length)
}
\arguments{
\item{indel_rep}{an integer value}

\item{is_length_ok}{an booleen value}

\item{diff_mutation_bam}{an integer value}

\item{deletion_length}{an integer value}
}
\value{
0 or 1 if the indel is the one we're looking for
}
\description{
Check if the indel is in a repeted sequence
}
\keyword{internal}
