% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_heatmap.R
\name{cohort_heatmap}
\alias{cohort_heatmap}
\title{Cohort heatmap
This function creates a heatmap for the cohort data in a list of matrices}
\usage{
cohort_heatmap(
  matrix_data,
  signal = "central_coverage",
  figure_title = "Cohort heatmap"
)
}
\arguments{
\item{matrix_data}{A list of matrices containing the cohort data.}

\item{signal}{A character string specifying the
signal to use for the heatmap.}

\item{figure_title}{A character string specifying
the title of the list of heatmaps.}
}
\value{
A heatmap object.
}
\description{
Cohort heatmap
This function creates a heatmap for the cohort data in a list of matrices
}
