% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_matrixes.R
\name{cohort_matrixes}
\alias{cohort_matrixes}
\title{Create a list of matrices for each timepoint in the experiment}
\usage{
cohort_matrixes(experiment, signal = "central_coverage")
}
\arguments{
\item{experiment}{A tibble containing the experiment data.}

\item{signal}{A character string specifying the signal to use for the matrix.}
}
\value{
A list of matrices, each corresponding to
a timepoint in the experiment.
}
\description{
Create a list of matrices for each timepoint in the experiment
}
