% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_base_basq_mstat_from_read.R
\name{compare_read_to_ref_wt_and_mut}
\alias{compare_read_to_ref_wt_and_mut}
\title{Get mutation status of read by comparing to wild-type and mutated reference}
\usage{
compare_read_to_ref_wt_and_mut(
  read_seq,
  ref_seq_wt,
  ref_seq_mut,
  compare_len_wt,
  compare_len_mut
)
}
\arguments{
\item{read_seq}{Base sequence of the read}

\item{ref_seq_wt}{Base sequence of the wild-type ref}

\item{ref_seq_mut}{Base sequence of the mutated ref}

\item{compare_len_wt}{Number of bases to be compared in read vs wild-type ref comparison}

\item{compare_len_mut}{Number of bases to be compared in read vs mutated ref comparison}
}
\value{
a character describing the mutation status of the read
}
\description{
Get mutation status of read by comparing to wild-type and mutated reference
}
\keyword{internal}
