% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{custom_legend}
\alias{custom_legend}
\title{custom_legend
Function to create a custom legend for ggplot2.
NOTE: the labels are reorder in alphabetical order.}
\usage{
custom_legend(
  labels = NULL,
  ...,
  title = NULL,
  key = ggplot2::draw_key_point,
  guide_args = list()
)
}
\arguments{
\item{labels}{A character vector of labels for the legend.}

\item{...}{A variable number of aesthetic mappings to be used in the legend.}

\item{title}{A character string for the title of the legend.}

\item{key}{A function to draw the key glyph
in the legend (default is \code{draw_key_point}).}

\item{guide_args}{A list of additional arguments
to be passed to the \code{guide_legend} function.}
}
\value{
A list containing a dummy geom and a dummy scale for the legend.
}
\description{
custom_legend
Function to create a custom legend for ggplot2.
NOTE: the labels are reorder in alphabetical order.
}
