% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{define_ins_rep}
\alias{define_ins_rep}
\title{Define the number of times the insertion is repeted}
\usage{
define_ins_rep(alt_len, current_pos, r_pos, cigar, r_query)
}
\arguments{
\item{alt_len}{an integer value representing the length of the insertion}

\item{current_pos}{an integer value}

\item{r_pos}{numeric value representing the read mapping position}

\item{cigar}{character vector representing the read CIGAR}

\item{r_query}{character vector representing the read base sequence}
}
\value{
an integer value which represents the number of the sequence insertion that is repeted in the sequence
}
\description{
Return an integer value which represents the number of the sequence insertion that is repeted in the sequence
}
\keyword{internal}
