% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-define-mutation-status.R
\name{define_mutation_status}
\alias{define_mutation_status}
\title{Define Mutation Status
Determines the type of mutation (SNV, MNP, insertion, or deletion) based on
the length of the REF and ALT strings.}
\usage{
define_mutation_status(ref, alt)
}
\arguments{
\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}
}
\value{
A character string indicating the type of mutation:
one of "SNV", "MNP", "insertion", or "deletion".
}
\description{
Define Mutation Status
Determines the type of mutation (SNV, MNP, insertion, or deletion) based on
the length of the REF and ALT strings.
}
\keyword{internal}
