% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-extract_fragment_features.R
\name{extract_fragment_features}
\alias{extract_fragment_features}
\title{Process a single sequencing fragment
Extracts and processes relevant data from a sequencing fragment.}
\usage{
extract_fragment_features(
  df_sam,
  fragment_name,
  sample_id,
  chr,
  pos,
  ref,
  alt,
  report_tlen,
  report_softclip,
  report_5p_3p_bases_fragment,
  cigar_free_indel_match,
  fasta_fafile = NULL,
  fasta_seq = NULL
)
}
\arguments{
\item{df_sam}{A dataframe containing sequencing reads.}

\item{fragment_name}{Name of the fragment (paired-end reads).}

\item{sample_id}{Sample identifier.}

\item{chr}{Character vector representing the chromosome of interest.}

\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{report_tlen}{Boolean. Whether to include the TLEN (template length)
information in the output.}

\item{report_softclip}{Boolean. Whether to include the number of soft-clipped
bases at the fragment extremities in the output.}

\item{report_5p_3p_bases_fragment}{Integer. Whether to include N fragment
extremity bases in the output.}

\item{cigar_free_indel_match}{Boolean. If activated, the information from the CIGAR is disregarded when determining the
mutation status of a read for indel. Instead the mutation status is determined by comparing the sequence of the read
to the sequence of the wild-type reference and the mutated reference. Activating this option may lead to discordant
genotyping of reads compared to the information provided by the CIGAR for indels. On the other hand, when
activated, it may rescue mutated genotypes for indel that would be missed in cases where the representation of the
indel in the CIGAR does not match the norm of bcftools of the mutation being analyzed.}

\item{fasta_fafile}{An open connection to an object of class FaFile}

\item{fasta_seq}{A list with the fasta sequence between two positions.}
}
\value{
A dataframe with the processed fragment information.
}
\description{
Process a single sequencing fragment
Extracts and processes relevant data from a sequencing fragment.
}
\keyword{internal}
