% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_c_qual}
\alias{find_c_qual}
\title{Find quality of the base}
\usage{
find_c_qual(alt_len, current_pos, r_pos, r_cigar, r_qual)
}
\arguments{
\item{alt_len}{an integer value representing the length of the insertion}

\item{current_pos}{an integer value representing the current position of the insertion before the cigar operations}

\item{r_pos}{numeric value representing the read mapping position}

\item{r_qual}{character vector representing the read sequencing qualities}

\item{cigar}{character vector representing the read CIGAR}
}
\value{
0 or 1 if the indel is the one we're looking for
}
\description{
Find quality of the base
}
\keyword{internal}
