% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRagmentomics-package.R
\docType{package}
\name{fRagmentomics-package}
\alias{fRagmentomics}
\alias{fRagmentomics-package}
\title{fRagmentomics: Extract Fragmentomics Features and Mutational Status}
\description{
A user-friendly R package that enables the characterization of each cfDNA fragment overlapping one or multiple mutations of interest, starting from a sequencing file containing aligned reads (BAM file). fRagmentomics supports multiple mutation input formats (e.g., VCF, TSV, or string "chr:pos:ref:alt" representation), accommodates one-based and zero-based genomic conventions, handles mutation representation ambiguities, and accepts any reference file and species in FASTA format. For each cfDNA fragment, fRagmentomics outputs its size, its 3' and 5' sequences, and its mutational status.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ElsaB-Lab/fRagmentomics}
  \item Report bugs at \url{https://github.com/ElsaB-Lab/fRagmentomics/issues}
}

}
\author{
\strong{Maintainer}: Killian Maudet \email{killian.maudet@gustaveroussy.fr} (\href{https://orcid.org/0009-0003-3237-092X}{ORCID})

Authors:
\itemize{
  \item Yoann Pradat \email{yoann.pradat@gustaveroussy.fr} (\href{https://orcid.org/0000-0002-4647-5779}{ORCID})
  \item Juliette Samaniego \email{juliette.samaniego@gustaveroussy.fr}
  \item Elsa Bernard \email{elsa.bernard@gustaveroussy.fr} (\href{https://orcid.org/0000-0002-2057-7187}{ORCID})
}

}
\keyword{internal}
