% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRagmentomics-package.R
\name{fRagmentomics}
\alias{fRagmentomics}
\title{fRagmentomics: Characterization of cfDNA Fragments}
\description{
A user-friendly R package that enables the characterization
of each cfDNA fragment overlapping one or multiple mutations of interest,
starting from a sequencing file containing aligned reads (BAM file).
fRagmentomics supports multiple mutation input formats (e.g., VCF, TSV,
or string “chr:pos:ref:alt” representation), accommodates one-based and
zero-based genomic conventions, handles mutation representation ambiguities,
and accepts any reference file and species in FASTA format. For each cfDNA
fragment, fRagmentomics outputs its size, its 3’ and 5’ sequences, and its
mutational status.
}
\examples{
# Example usage of fRagmentomics
# Load the package
library(fRagmentomics)

# Example function call
# result <- analyze_fragments(bam_file = "path/to/your/file.bam",
#                             mutation_file = "path/to/your/mutations.vcf")

}
\author{
Killian Maudet \href{mailto:killian.maudet@gustaveroussy.fr}{killian.maudet@gustaveroussy.fr} \link{aut, cre}

Yoann Pradat \href{mailto:yoann.pradat@gustaveroussy.fr}{yoann.pradat@gustaveroussy.fr} \link{aut, cre}

Juliette Samaniego \href{mailto:juliette.samaniego@gustaveroussy.fr}{juliette.samaniego@gustaveroussy.fr} \link{aut}

Elsa Bernard \href{mailto:elsa.bernard@gustaveroussy.fr}{elsa.bernard@gustaveroussy.fr} \link{aut}
}
