% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_base_basq_mstat_from_read.R
\name{get_base_basq_from_read_at_pos}
\alias{get_base_basq_from_read_at_pos}
\title{Get the element in the sequence and quality aligning with the position of interest.}
\usage{
get_base_basq_from_read_at_pos(pos_cur, pos_pre, read_stats)
}
\arguments{
\item{pos_cur}{current position request}

\item{pos_pre}{previous position request}

\item{read_stats}{A list of read-level statistics.}
}
\value{
A list with the base and the quality of the read aligning with the position provided. If the read contains a
deletion, base is set to '-' and the quality to an empty string.
}
\description{
Extracts base, quality, and indel informations based on the mutation type.
}
\keyword{internal}
