% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_basq_mstat_from_read.R
\name{get_base_basq_mstat_from_read}
\alias{get_base_basq_mstat_from_read}
\title{Extract allele, quality, and mutation status from a read}
\usage{
get_base_basq_mstat_from_read(
  chr,
  pos,
  ref,
  alt,
  read_stats,
  fasta_fafile = NULL,
  fasta_seq = NULL
)
}
\arguments{
\item{chr}{Character vector representing the chromosome of interest.}

\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{read_stats}{A list of read-level statistics.}

\item{fasta_fafile}{An open connection to an object of class FaFile}

\item{fasta_seq}{A list with the fasta sequence between two positions.}
}
\value{
A list containing 'base', 'basq', 'mstat'.
}
\description{
This function serves as a primary parser for a single read at a
variant locus. It determines if the read covers the variant, extracts the
observed sequence and its base qualities, and assigns a detailed mutation
status.
}
\keyword{internal}
