% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector-deletion.R
\name{get_deletion}
\alias{get_deletion}
\title{Get the informations about the presence of a deletion in the read.}
\usage{
get_deletion(pos, ref, r_pos, r_cigar, r_qual)
}
\arguments{
\item{pos}{numeric value representing the position of interest}

\item{ref}{character vector representing the deletion sequence}

\item{r_pos}{numeric value representing the read mapping position}

\item{r_cigar}{character vector representing the read CIGAR}

\item{r_qual}{character vector representing the read sequencing qualities}
}
\value{
a named list with names \code{base},\code{qual}.
}
\description{
Get the informations about the presence of a deletion in the read.
}
\keyword{internal}
