% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_fragment_bases_5p_3p.R
\name{get_fragment_bases_5p_3p}
\alias{get_fragment_bases_5p_3p}
\title{Extract Fragment Bases and Quality Scores from Read Sequences}
\usage{
get_fragment_bases_5p_3p(n_bases, seq_5p, seq_3p, qual_5p, qual_3p)
}
\arguments{
\item{n_bases}{The number of bases to extract from each end (5' and 3').}

\item{seq_5p}{The nucleotide sequence of the 5' read.}

\item{seq_3p}{The nucleotide sequence of the 3' read.}

\item{qual_5p}{The quality string corresponding to the 5' read.}

\item{qual_3p}{The quality string corresponding to the 3' read.}
}
\value{
A list containing four character elements:
-fragment_bases_5p: The first 'n_bases' nucleotides from the 5' read.
-fragment_bases_3p: The last 'n_bases' nucleotides from the 3' read.
-fragment_basqs_5p: The first 'n_bases' qual chr from the 5' read.
-fragment_basqs_3p: The last 'n_bases' qual chr from the 3' read.
}
\description{
This function extracts a fixed number of bases ('n_bases')
and their corresponding quality scores from the 5' and 3' ends of reads.
}
\keyword{internal}
