% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_fragment_inner_distance.R
\name{get_fragment_inner_distance}
\alias{get_fragment_inner_distance}
\title{Compute fragment inner distance
Calculates the inner distance between two paired-end sequencing reads.}
\usage{
get_fragment_inner_distance(r_pos_5p, r_cigar_5p, r_pos_3p, r_cigar_3p)
}
\arguments{
\item{r_pos_5p}{numeric value representing the read 5p mapping position}

\item{r_cigar_5p}{character vector representing the read 5p CIGAR}

\item{r_pos_3p}{numeric value representing the read 3p mapping position}

\item{r_cigar_3p}{character vector representing the read 3p CIGAR}
}
\value{
a named list with names \code{a} and \code{b}.
}
\description{
Compute fragment inner distance
Calculates the inner distance between two paired-end sequencing reads.
}
\keyword{internal}
