% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_fragment_mutation_status.R
\name{get_fragment_mutation_status_broad}
\alias{get_fragment_mutation_status_broad}
\title{Get fragment mutation status according to the mutation status of each read.
Detailed label of get_fragment_mutation_status}
\usage{
get_fragment_mutation_status_broad(mstat_1, mstat_2)
}
\arguments{
\item{mstat_1}{The mutation status of read 1}

\item{mstat_2}{The mutation status of read 2}
}
\value{
A character string representing the fragment status:
}
\description{
This function determines the mutation status of a DNA fragment based on
the mutation status of each read from the fragment.
}
\keyword{internal}
