% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_fragment_mutation_statuses.R
\name{get_fragment_mutation_statuses}
\alias{get_fragment_mutation_statuses}
\title{Get detailed and simplified fragment mutation statuses.}
\usage{
get_fragment_mutation_statuses(mstat_1, mstat_2)
}
\arguments{
\item{mstat_1}{The mutation status of read 1 (e.g., "MUT:C>T_high_conf", "WT", "AMB_low_coverage").
Can be \code{NA} for no coverage.}

\item{mstat_2}{The mutation status of read 2 (e.g., "MUT:C>T_high_conf", "WT", "AMB_low_coverage").
Can be \code{NA} for no coverage.}
}
\value{
A list containing two character strings:
\itemize{
\item \code{Fragment_Status_Detail}: A detailed status reflecting the combination,
retaining original text where applicable (e.g., "MUT & AMB_low_coverage").
\item \code{Fragment_Status_Simple}: A simplified status (e.g., "MUT", "NON-TARGET MUT",
"DISCORDANT", "AMB", "ERROR").
}
}
\description{
This function determines both the detailed and simplified mutation statuses of a DNA
fragment based on the mutation status of each of the two reads from the fragment.
}
\keyword{internal}
