% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_fragment_size.R
\name{get_fragment_size}
\alias{get_fragment_size}
\title{Compute fragment size
Calculates the fragment size with two paired-end sequencing reads.
Taking into account soft clipping, insertion, deletion}
\usage{
get_fragment_size(read_stats_5p, read_stats_3p)
}
\arguments{
\item{read_stats_5p}{5p read infos}

\item{read_stats_3p}{3p read infos}
}
\value{
a integer corresponding to the fragment size
}
\description{
Compute fragment size
Calculates the fragment size with two paired-end sequencing reads.
Taking into account soft clipping, insertion, deletion
}
\keyword{internal}
