% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector-deletion.R
\name{get_info_deletion}
\alias{get_info_deletion}
\title{Get the informations about the presence of a deletion in the read.}
\usage{
get_info_deletion(pos, ref, r_pos, r_cigar, r_qual, pos_after_indel_repetition)
}
\arguments{
\item{pos}{numeric value representing the position of interest}

\item{ref}{character vector representing the VCF REF allele for the deletion.}

\item{r_pos}{numeric value representing the read mapping position}

\item{r_cigar}{character vector representing the read CIGAR}

\item{r_qual}{character vector representing the read sequencing qualities}

\item{pos_after_indel_repetition}{integer, 1-based genomic position of the
first reference nucleotide after the (potentially repetitive) region involved
in the indel. This position is used to determine if a read not showing the
deletion is long enough to be called Wild Type or if it's ambiguous.}
}
\value{
a named list with names 'base','qual'.
'base' can be:
\itemize{
\item "no_deletion_detected": Read covers the site, shows no deletion, and is
long enough to be unambiguous = WT.
\item "-DELETED_SEQ": The specific deletion was detected (e.g., "-ATC").
\item NA_character : Read does not cover the deletion site.
\item "ambiguous": Read coverage is insufficient relative to 'pos_after_indel_repetition'
to make an unambiguous call. This status is assigned if
'last_ref_pos_covered_by_read < (pos_after_indel_repetition - deletion_length)'.
'qual' provides the quality score of the base preceding the deletion if
detected, "ambiguous" if base is "ambiguous", NA if base is
NA, or "no_deletion_detected".
}
}
\description{
Get the informations about the presence of a deletion in the read.
}
\keyword{internal}
