% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get-mutation_info.R
\name{get_mutation_info}
\alias{get_mutation_info}
\title{Retrieve mutation-specific information:
Extracts base, quality, and indel informations based on the mutation type.}
\usage{
get_mutation_info(
  mutation_type,
  pos,
  ref,
  alt,
  read_stats,
  pos_after_indel_repetition
)
}
\arguments{
\item{mutation_type}{In "SNV", "deletion", or "insertion".}

\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{read_stats}{A list of read-level statistics.}

\item{pos_after_indel_repetition}{Integer. The position after the indel
repetion.}
}
\value{
A list containing "base", "qual".
Base could be:
\itemize{
\item NA if the read doesn't cover the position of interest
\item Nucleodide for SNV if the read covers the position of interest
\item The sequence of the deletion or insertion if detected
\item "no_deletion_detected" or "no_insertion_detected"
}

Qual could be:
\itemize{
\item NA if the read doesn't cover the position of interest
\item Qual for SNV if the read covers the position of interest
\item Qual of nucleotide before the indel if detected
\item "no_deletion_detected" or "no_insertion_detected"
}
}
\description{
Retrieve mutation-specific information:
Extracts base, quality, and indel informations based on the mutation type.
}
\keyword{internal}
