% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_alt_basq_mstat_from_read.R
\name{get_nb_indices_to_report}
\alias{get_nb_indices_to_report}
\title{Get number of indices to be used for reporting}
\usage{
get_nb_indices_to_report(ref, alt)
}
\arguments{
\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}
}
\value{
A number indicate the number of nucleotides for which stats will be reported
}
\description{
Get number of indices to be used for reporting
}
\keyword{internal}
