% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_alt_basq_mstat_from_read.R
\name{get_number_of_common_first_char}
\alias{get_number_of_common_first_char}
\title{Get the number of common first characters beetween two strings}
\usage{
get_number_of_common_first_char(str_a, str_b)
}
\arguments{
\item{str_a}{a string}

\item{str_b}{a string}
}
\value{
An integer
}
\description{
Get the number of common first characters beetween two strings
}
\keyword{internal}
