% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector-insertion.R
\name{get_seq_ins}
\alias{get_seq_ins}
\title{sanity_check_ins}
\usage{
get_seq_ins(alt_len, current_pos, r_pos, cigar, r_query)
}
\arguments{
\item{alt_len}{an integer value representing the length of the insertion}

\item{current_pos}{an integer value}

\item{r_pos}{numeric value representing the read mapping position}

\item{cigar}{character vector representing the read CIGAR}

\item{r_query}{character vector representing the read base sequence}
}
\value{
a character vector representing the insertion sequence
}
\description{
sanity_check_ins
}
\keyword{internal}
