% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample_data.R
\name{group_targets}
\alias{group_targets}
\title{Group targets data}
\usage{
group_targets(samples, samplesheet)
}
\arguments{
\item{samples}{A list of samples matrices loaded from the samplesheet.}

\item{samplesheet}{A tibble containing the samplesheet data.}
}
\description{
Group targets data
}
