% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{heatmap_color}
\alias{heatmap_color}
\title{Returns a color function for the combined matrix_data}
\usage{
heatmap_color(matrix_data, colors = c("red", "white", "blue"))
}
\arguments{
\item{matrix_data}{A list of matrices to be used for color scaling.}

\item{colors}{A character vector of 3 colors to be used in the heatmap.
default: red, white, blue with red for negative values}
}
\value{
A color function that can be used for heatmaps.
}
\description{
Returns a color function for the combined matrix_data
}
