% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_plot.R
\name{housekeeping_plot_relative}
\alias{housekeeping_plot_relative}
\title{Plot Relative Housekeeping Genes}
\usage{
housekeeping_plot_relative(housekeeping, random)
}
\arguments{
\item{housekeeping}{A tibble containing the housekeeping gene peak data.}

\item{random}{A tibble containing the random datasets peak data.}
}
\description{
Plot Relative Housekeeping Genes
}
