% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_zscore.R
\name{housekeeping_zscore}
\alias{housekeeping_zscore}
\title{Zscore for housekeeping genes}
\usage{
housekeeping_zscore(housekeeping, random)
}
\arguments{
\item{housekeeping}{A tibble containing the housekeeping gene stats data.}

\item{random}{A tibble containing the random datasets stats data.}
}
\value{
A tibble containing the z-scores for the housekeeping genes.
}
\description{
This function calculates the z-score for housekeeping
genes compared to random datasets.
}
