% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_zscore_plot.R
\name{housekeeping_zscore_plot}
\alias{housekeeping_zscore_plot}
\title{Housekeeping Z-score Plot}
\usage{
housekeeping_zscore_plot(housekeeping, random, bins = NULL)
}
\arguments{
\item{housekeeping}{A tibble containing the housekeeping gene stats data.}

\item{random}{A tibble containing the random datasets stats data.}

\item{bins}{An optional integer specifying
the number of bins for the histogram.
If NULL, breaks are calculated using Sturge's method.}
}
\value{
A ggplot object for the z-scores of housekeeping genes.
}
\description{
This function generates a ggplot plot for the z-scores of housekeeping genes
compared to random datasets.
}
