% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot.R
\name{label_basepairs}
\alias{label_basepairs}
\title{Label basepairs}
\usage{
label_basepairs(x)
}
\arguments{
\item{x}{A numeric vector representing basepairs.}
}
\value{
A character vector with basepairs labeled with appropriate suffixes.
}
\description{
Label basepairs
}
