% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experiment.R
\name{load_experiment}
\alias{load_experiment}
\title{Load sample data from a samplesheet
This function is intended to load sample data
based on a provided samplesheet.}
\usage{
load_experiment(
  samplesheet,
  rootpath,
  subdir = "fragmentomics/processed",
  parallelize = TRUE,
  skip_matrix_files = FALSE,
  number_of_daemons = parallelly::availableCores()
)
}
\arguments{
\item{samplesheet}{A character string specifying
the path to the samplesheet file.}

\item{rootpath}{A character string specifying the
root path where the sample data is located.}

\item{subdir}{A character string specifying the
subdirectory within the root path where
the sample data is located, defaults
to "fragmentomics/processed/matrix".}

\item{parallelize}{A logical value indicating whether to parallelize the
loading of sample data, defaults to TRUE.#' loading of sample data,
defaults to TRUE.}

\item{skip_matrix_files}{A logical value indicating whether to skip loading
matrix files, defaults to FALSE.}

\item{number_of_daemons}{An integer specifying the number of daemons to use}
}
\value{
A tibble containing the loaded sample data.
}
\description{
Load sample data from a samplesheet
This function is intended to load sample data
based on a provided samplesheet.
}
\examples{
\dontrun{
example_samplesheet <- system.file("extdata",
                                   "samplesheet.csv",
                                   package = "fragmentomics")
samplesheet <- parse_samplesheet(example_samplesheet)
experiment <- load_experiment(samplesheet, "results")
}
}
