% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample_data.R
\name{load_matrix_files}
\alias{load_matrix_files}
\title{Load matrix files from a directory}
\usage{
load_matrix_files(
  caseid,
  sampleid,
  timepoint,
  encoded_timepoint,
  rootpath,
  subdir
)
}
\arguments{
\item{caseid}{A character string specifying the case ID.}

\item{sampleid}{A character string specifying the sample ID.}

\item{timepoint}{A factor specifying the timepoint.}

\item{encoded_timepoint}{An integer specifying the encoded timepoint.}

\item{rootpath}{A character string specifying the root path where the sample data is located.}

\item{subdir}{A character string specifying the subdirectory within the root path where
the sample data is located, defaults to "fragmentomics/processed/matrix".}
}
\description{
This function is intended to load matrix files from a specified directory.
}
