% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_peaks.R
\name{load_peak_data}
\alias{load_peak_data}
\title{load_peak_data
Load peak data from a specified file and add metadata.}
\usage{
load_peak_data(
  caseid,
  sampleid,
  timepoint,
  encoded_timepoint,
  signal_label,
  target_label,
  source_label,
  bin_size,
  central_bin,
  peakfile
)
}
\arguments{
\item{caseid}{A character string representing the case ID.}

\item{sampleid}{A character string representing the sample ID.}

\item{timepoint}{A factor representing the timepoint.}

\item{encoded_timepoint}{An integer representing the encoded timepoint.}

\item{signal_label}{A character string representing the signal label.}

\item{target_label}{A character string representing the target label.}

\item{source_label}{A character string representing the source label.}

\item{bin_size}{An integer representing the size of
the bins used in the peak data.}

\item{central_bin}{An integer representing the central bin of the peak data.}

\item{peakfile}{A character string representing
the path to the peak data file.}
}
\value{
A tibble containing the parsed peak data with additional metadata.
}
\description{
load_peak_data
Load peak data from a specified file and add metadata.
}
