% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_peaks.R
\name{load_peaks}
\alias{load_peaks}
\title{Load peaks
take in input an experiment data structure and return
a peaks data structure loading peaks data files}
\usage{
load_peaks(
  experiment,
  rootpath,
  subdir = "fragmentomics/processed",
  parallelize = TRUE,
  number_of_daemons = parallel::detectCores()
)
}
\arguments{
\item{experiment}{A tibble containing the experiment data.}

\item{rootpath}{A character string specifying the
root path where the peaks data is located.}

\item{subdir}{A character string specifying the subdirectory within the
root path where the peaks data is located,
defaults to "fragmentomics/processed".}

\item{parallelize}{A logical value indicating whether to
parallelize the loading of peaks data, defaults to TRUE.}

\item{number_of_daemons}{An integer specifying the number of
daemons to use for parallel processing,
defaults to the number of available cores.}
}
\value{
A tibble containing the loaded peaks data.
}
\description{
Load peaks
take in input an experiment data structure and return
a peaks data structure loading peaks data files
}
\examples{
\dontrun{
example_samplesheet <- system.file("extdata",
                                   "samplesheet.csv",
                                   package = "fragmentomics")
samplesheet <- parse_samplesheet(example_samplesheet)
experiment <- load_experiment(samplesheet, "results")
ctcf <- experiment |> dplyr::filter(target_label == "CTCF")
ctcf_peaks <- load_peaks(ctcf, "results/")
}
}
