% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_mut.R
\name{normalize_mut}
\alias{normalize_mut}
\title{Normalize a data frame of variants}
\usage{
normalize_mut(df_mut, fasta, fasta_fafile, one_based, tmp_folder)
}
\arguments{
\item{df_mut}{a dataframe with mutations}

\item{fasta}{Path to the FASTA file for the reference sequence used for generating the BAM file.}

\item{fasta_fafile}{An open connection to an object of class FaFile}

\item{one_based}{Boolean. TRUE if fasta is in one based. False if in 0 based.}

\item{tmp_folder}{Crun_fRagmentomicsharacter vector for the temporary folder path.}
}
\value{
a dataframe with normalized mutations representations containing columns CHROM, POS, REF, ALT
}
\description{
This function serves as the main variant normalization pipeline. It iterates through a data frame of variants,
applying a two-step normalization process to each one, ensuring they are represented in a canonical, left-aligned
format suitable for downstream analysis.
}
\keyword{internal}
