% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-normalize_to_vcf_rep.R
\name{normalize_to_vcf_rep}
\alias{normalize_to_vcf_rep}
\title{Normalize user-provided representation into VCF representation
This function normalizes the REF and ALT columns into the VCF format.
It also checks if all nucleotides in the REF column match those from
a given FASTA reference.}
\usage{
normalize_to_vcf_rep(chr, pos, ref, alt, fasta_fafile, one_based)
}
\arguments{
\item{chr}{A string representing the chromosome.}

\item{pos}{An integer representing the position.}

\item{ref}{A string representing the reference allele.}

\item{alt}{A string representing the alternative allele.}

\item{fasta_fafile}{An open connection to an object of class FaFile}

\item{one_based}{Boolean. TRUE if fasta is in one based. False if in 0 based.}
}
\value{
A list with chr, pos, ref and alt
}
\description{
Normalize user-provided representation into VCF representation
This function normalizes the REF and ALT columns into the VCF format.
It also checks if all nucleotides in the REF column match those from
a given FASTA reference.
}
\keyword{internal}
