% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-user-to-vcf-repr.R
\name{normalize_user_rep_to_vcf_rep}
\alias{normalize_user_rep_to_vcf_rep}
\title{Normalize user-provided representation into VCF representation
This function normalizes the REF and ALT columns into the VCF format.
It also checks if all nucleotides in the REF column match those from
a given FASTA reference.}
\usage{
normalize_user_rep_to_vcf_rep(chr, pos, ref, alt, fasta, one_based)
}
\arguments{
\item{chr}{Character vector representing the chromosome of interest.}

\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{fasta}{A reference genome in FASTA format.}

\item{one_based}{Boolean. TRUE if fasta is in one based. False if in 0 based.}
}
\value{
A list with chr, pos, ref and alt
}
\description{
Normalize user-provided representation into VCF representation
This function normalizes the REF and ALT columns into the VCF format.
It also checks if all nucleotides in the REF column match those from
a given FASTA reference.
}
\keyword{internal}
