% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot.R
\name{normalized_plot}
\alias{normalized_plot}
\title{Normalized plot}
\usage{
normalized_plot(data, palette)
}
\arguments{
\item{data}{A list containing the average data and stats.}

\item{palette}{A character vector representing the color
palette to be used for the plot.A character vector
representing the color palette to be used for the plot.}
}
\value{
A ggplot object for normalized data.
}
\description{
This function generates a ggplot plot for normalized data.
}
