% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_peaks.R
\name{parallel_load_peaks}
\alias{parallel_load_peaks}
\title{parallel_load_peaks
Load peaks data in parallel.}
\usage{
parallel_load_peaks(mdata)
}
\arguments{
\item{mdata}{A tibble containing the metadata for the peaks data.}
}
\value{
A tibble containing the loaded peaks data.
}
\description{
parallel_load_peaks
Load peaks data in parallel.
}
