% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experiment.R
\name{parallel_load_samples}
\alias{parallel_load_samples}
\title{Load samples in parallel
This function is intended to load samples in parallel}
\usage{
parallel_load_samples(samplesheet, rootpath, subdir, skip_matrix_files)
}
\arguments{
\item{samplesheet}{A tibble containing the samplesheet data.}

\item{rootpath}{A character string specifying the
root path where the sample data is located.}

\item{subdir}{A character string specifying the}

\item{skip_matrix_files}{A logical value indicating whether to skip loading
subdirectory within the root path where
the sample data is located, defaults
to "fragmentomics/processed/matrix".}
}
\value{
A tibble containing the loaded sample data.
}
\description{
Load samples in parallel
This function is intended to load samples in parallel
}
