% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_cigar}
\alias{parse_cigar}
\title{parse_cigar}
\usage{
parse_cigar(cigar)
}
\arguments{
\item{cigar}{a character vector}
}
\value{
a df with length and type of the CIGAR string. One row per operation.
}
\description{
Function to parse the CIGAR string
}
\keyword{internal}
