% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_compute_matrix.R
\name{parse_compute_matrix}
\alias{parse_compute_matrix}
\title{Parse Compute Matrix}
\usage{
parse_compute_matrix(file)
}
\arguments{
\item{file}{Path to the computeMatrix output file.}
}
\value{
A tibble containing the parsed data.
\enumerate{
\item columns 1-23 are metadata columns
\item columns 24-29 are chr, start, end, name, score, strand
\item columns 30+ are bins with values
}
}
\description{
This function parse output from deeptools computeMatrix
https://deeptools.readthedocs.io/en/develop/content/tools/computeMatrix.html
}
\examples{
file <- system.file("extdata", "CTCF_matrix.gz", package = "fragmentomics")
matrix_data <- parse_compute_matrix(file)
}
