% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_peak_data.R
\name{parse_peak_data}
\alias{parse_peak_data}
\title{Parse peak data}
\usage{
parse_peak_data(file)
}
\arguments{
\item{file}{A character string specifying the path to the peak data file.}
}
\description{
This function reads peak data from a file and returns it as a tibble.
}
