% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_peak_stats.R
\name{parse_peak_stats}
\alias{parse_peak_stats}
\title{Parse peak stats from a file}
\usage{
parse_peak_stats(file)
}
\arguments{
\item{file}{Path to the file containing peak stats.}
}
\value{
A tibble with the parsed peak stats.
}
\description{
Parse peak stats from a file
}
\examples{
file <- system.file("extdata", "CTCF_peak_stats.tsv", package = "fragmentomics")
parse_peak_stats(file)
}
