% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_samplesheet.R
\name{parse_samplesheet}
\alias{parse_samplesheet}
\title{Parse a samplesheet file}
\usage{
parse_samplesheet(file, root_path, timepoint_levels = c("BL", "BR", "PD"))
}
\arguments{
\item{file}{A character string specifying the path to the samplesheet file.}

\item{root_path}{A character string specifying the root
path for the samplesheet.}

\item{timepoint_levels}{A character vector specifying the levels for the
timepoint factor.}
}
\value{
A tibble containing the parsed samplesheet data.
}
\description{
This function is intended to parse a samplesheet file,
which typically contains metadata about samples
used in an experiment.
}
\examples{
example_samplesheet <- system.file(
  "extdata",
  "samplesheet.csv",
  package = "fragmentomics"
)
parse_samplesheet(example_samplesheet)

}
