% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseComputeMatrix.R
\name{parseComputeMatrix}
\alias{parseComputeMatrix}
\title{Parse Compute Matrix}
\usage{
parseComputeMatrix(file)
}
\arguments{
\item{file}{Path to the computeMatrix output file.}
}
\value{
A tibble containing the parsed data.
\enumerate{
\item columns 1-23 are metadata columns
\item columns 24-29 are chr, start, end, name, score, strand
\item columns 30+ are bins with values
}
}
\description{
This function parse output from deeptools \href{https://deeptools.readthedocs.io/en/develop/content/tools/computeMatrix.html}{computeMatrix}
}
