% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot.R
\name{peak_length}
\alias{peak_length}
\title{peak_length}
\usage{
peak_length(data, normalized = FALSE)
}
\arguments{
\item{data}{A data frame containing the average data and stats.}

\item{normalized}{A logical value indicating whether the data is normalized.}
}
\value{
A tibble representing the peak length.
}
\description{
Function to calculate the peak length based on the data.
}
