% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_stats.R
\name{peak_stats}
\alias{peak_stats}
\title{Calculate background mean using the limits of the distribution.
Aka the extreme parts of the peak that will represent the background}
\usage{
peak_stats(
  data,
  signal_label = NULL,
  target_label = NULL,
  source_label = NULL,
  left = 50,
  right = 50,
  central = 30,
  average = 1000,
  rpoints = 10000
)
}
\arguments{
\item{data}{A tibble with the coverage data}

\item{signal_label}{A label for the signal, defaults to sample_labels}

\item{target_label}{A label for the target, defaults to file}

\item{source_label}{A label for the source, defaults to NA}

\item{left}{The number of bins to the left of the peak
to use for background calculation, defaults to 50}

\item{right}{The number of bins to the right of the peak
to use for background calculation, defaults to 50}

\item{central}{The number of bins around the central point
to use for peak calculation, defaults to 30}

\item{average}{The number of bins around the central point
to use for average calculation, defaults to 1000}

\item{rpoints}{The number of random points to use for
Monte Carlo integration, defaults to 10000}
}
\value{
A list containing the peak statistics,
integration data, and average data
}
\description{
Calculate background mean using the limits of the distribution.
Aka the extreme parts of the peak that will represent the background
}
\examples{
file <- system.file("extdata", "CTCF.gz", package = "fragmentomics")
matrix_data <- parse_compute_matrix(file)
peak_data <- peak_stats(matrix_data)

}
