% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPlot.R
\name{peakPlot}
\alias{peakPlot}
\title{Peak Plot}
\usage{
peakPlot(data, normalized = FALSE)
}
\arguments{
\item{data}{A data frame containing the data to be plotted.}

\item{normalized}{A logical value indicating whether the data is normalized.
If TRUE, the plot will show relative signal;
if FALSE, it will show raw signal.}
}
\value{
A ggplot object.
}
\description{
This function generates a ggplot plot for the given data.
}
