% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot a matrix of coverage data}
\usage{
plot_matrix(matrix_data, palette = "RdYlBu")
}
\arguments{
\item{matrix_data}{A data frame containing the matrix data to be plotted.}

\item{palette}{A character string representing the color palette to be used for the plot.}
}
\value{
A ComplexHeatmap object representing the plotted matrix.
}
\description{
This function plots a matrix of coverage data using ComplexHeatmap.
}
