% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_get_fragment_bases.R
\name{process_get_fragment_bases}
\alias{process_get_fragment_bases}
\title{Extract Fragment Bases and Quality Scores from Read Sequences
This function extracts a fixed number of bases (\code{nbr_bases})
and their corresponding quality scores from the 5' and 3' ends of reads.}
\usage{
process_get_fragment_bases(nbr_bases, query_5p, query_3p, qual_5p, qual_3p)
}
\arguments{
\item{nbr_bases}{The number of bases to extract from each end (5' and 3').}

\item{query_5p}{The nucleotide sequence of the 5' read.}

\item{query_3p}{The nucleotide sequence of the 3' read.}

\item{qual_5p}{The quality string corresponding to the 5' read.}

\item{qual_3p}{The quality string corresponding to the 3' read.}
}
\value{
A list containing four character elements:
-fragment_bases_5p: The first \code{nbr_bases} nucleotides from the 5' read.
-fragment_bases_3p: The last \code{nbr_bases} nucleotides from the 3' read.
-fragment_Qbases_5p: The first \code{nbr_bases} qual chr from the 5' read.
-fragment_Qbases_3p: The last \code{nbr_bases} qual chr from the 3' read.
}
\description{
Extract Fragment Bases and Quality Scores from Read Sequences
This function extracts a fixed number of bases (\code{nbr_bases})
and their corresponding quality scores from the 5' and 3' ends of reads.
}
\keyword{internal}
