% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_get_fragment_bases_softclip.R
\name{process_get_fragment_bases_softclip}
\alias{process_get_fragment_bases_softclip}
\title{Extract Soft-Clipped Base Counts from CIGAR Strings
This function extracts the number of soft-clipped bases at
the 5' ends of 5' reads and 3' ends of 3' reads.
If no soft clipping is detected, the count is set to 0.}
\usage{
process_get_fragment_bases_softclip(cigar_5p, cigar_3p)
}
\arguments{
\item{cigar_5p}{Character string. The CIGAR string of the 5' read.}

\item{cigar_3p}{Character string. The CIGAR string of the 3' read.}
}
\value{
A list with two numeric elements:
-Nb_SC_5p: Number of soft-clipped bases at the 5' end.
-Nb_SC_3p: Number of soft-clipped bases at the 3' end.
}
\description{
Extract Soft-Clipped Base Counts from CIGAR Strings
This function extracts the number of soft-clipped bases at
the 5' ends of 5' reads and 3' ends of 3' reads.
If no soft clipping is detected, the count is set to 0.
}
\keyword{internal}
