% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot.R
\name{raw_plot}
\alias{raw_plot}
\title{Raw plot}
\usage{
raw_plot(data, palette)
}
\arguments{
\item{data}{A list containing the average data, integration data, and stats.}

\item{palette}{A character string representing the color palette
to be used for the plot.}
}
\value{
A ggplot object for raw data.
}
\description{
This function generates a ggplot plot for raw data.
}
