% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-read-bam.R
\name{read_bam}
\alias{read_bam}
\title{Preprocess BAM file to extract reads around a given position
This function extracts reads from a BAM file using Rsamtools, filtering them
based on flags and selecting reads of interest around a genomic position.}
\usage{
read_bam(bam, chr, pos, neg_offset, pos_offset, flag_keep, flag_remove)
}
\arguments{
\item{bam}{A dataframe containing sequencing reads.}

\item{chr}{Character. Chromosome of interest.}

\item{pos}{Integer. Genomic position of interest.}

\item{neg_offset}{Integer. Start position for wider read extraction.}

\item{pos_offset}{Integer. End position for wider read extraction.}

\item{flag_keep}{Integer. SAM flag for reads to keep.}

\item{flag_remove}{Integer. SAM flag for reads to exclude.}
}
\value{
A data frame containing the filtered SAM entries.
}
\description{
Preprocess BAM file to extract reads around a given position
This function extracts reads from a BAM file using Rsamtools, filtering them
based on flags and selecting reads of interest around a genomic position.
}
\keyword{internal}
