% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mut.R
\name{read_mut}
\alias{read_mut}
\title{Read variant information from multiple sources}
\usage{
read_mut(mut)
}
\arguments{
\item{mut}{Path to a .vcf or .tsv file or string representation chr:pos:ref:alt of a mutation.}
}
\value{
A 'data.frame' where each row represents a single bi-allelic variant, with the columns 'CHROM', 'POS', 'REF', and 'ALT'.
}
\description{
This function acts as a dispatcher to read variant information from a file (VCF or TSV) or a character
string. It automatically detects the input format and uses the appropriate parser. All multi-allelic sites are split
into separate, bi-allelic rows.
}
\keyword{internal}
