% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseComputeMatrix.R
\name{readComputeMatrix}
\alias{readComputeMatrix}
\title{Parse Compute Matrix}
\usage{
readComputeMatrix(file)
}
\arguments{
\item{file}{Path to the computeMatrix output file.}
}
\description{
This function parse output from deeptools computeMatrix
}
