% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-remove_bad_mut.R
\name{remove_bad_mut}
\alias{remove_bad_mut}
\title{Remove bad mutations}
\usage{
remove_bad_mut(df_mut)
}
\arguments{
\item{df_mut}{A dataframe with mutation information.}
}
\value{
A filtered dataframe containing only valid mutations.
}
\description{
This function verifies the validity of the mutation input data by
checking the chromosome, position, reference, and alternative alleles.
It only returns rows that pass all checks.
}
\keyword{internal}
