% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-accepted-ref-alt.R
\name{retrieve_anchor_base}
\alias{retrieve_anchor_base}
\title{Retrieve a single nucleotide from a multi-chromosome FASTA}
\usage{
retrieve_anchor_base(chr, position, fasta)
}
\arguments{
\item{chr}{A character string specifying the chromosome name "chr1".}

\item{position}{An integer specifying the genomic position (1-based).}

\item{fasta}{A FASTA reference that contains all chromosomes. Any reference object
compatible with Biostrings::getSeq.}
}
\value{
A character string representing the single nucleotide at the specified
chromosome and position.
}
\description{
This function retrieves a single nucleotide from a multi-chromosome FASTA reference
at the specified chromosome and position (1-based coordinates).
}
