% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanity_check_indel}
\alias{sanity_check_indel}
\title{Function to see verify if it's the indel that we're looking for}
\usage{
sanity_check_indel(is_length_ok, diff_mutation_bam, deletion_length)
}
\arguments{
\item{is_length_ok}{an booleen value}

\item{diff_mutation_bam}{an integer value}

\item{deletion_length}{an integer value}

\item{indel_rep}{an integer value}
}
\value{
0 or 1 if the indel is the one we're looking for
}
\description{
Function to see verify if it's the indel that we're looking for
}
\keyword{internal}
