% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-sanity-checks.R
\name{sanity_check_tsv}
\alias{sanity_check_tsv}
\title{Sanity Check TSV}
\usage{
sanity_check_tsv(tsv_file)
}
\arguments{
\item{tsv_file}{Path to the TSV file.}
}
\value{
A cleaned data frame with required columns.
}
\description{
This function checks if the TSV file contains the expected structure and columns.
}
