% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPlot.R
\name{scale_x_bins}
\alias{scale_x_bins}
\title{Function to build the x scale based on nother of bins}
\usage{
scale_x_bins(data)
}
\arguments{
\item{data}{A list containing the average data and stats.}
}
\value{
A list with breaks and labels for the x scale.
}
\description{
Function to build the x scale based on nother of bins
}
