% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_for_indel_in_cigar.R
\name{search_for_indel_in_cigar}
\alias{search_for_indel_in_cigar}
\title{Search for a specific indel within a read's CIGAR string}
\usage{
search_for_indel_in_cigar(pos, ref, alt, read_stats, type)
}
\arguments{
\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{read_stats}{A list of read-level statistics.}

\item{type, }{choose between 'INS' or 'DEL'}
}
\value{
a list with two booleans. The first boolean indicate if the INDEL searched was found. The second boolean
indicate if another INDEL (different size and/or different sequence) was found.
}
\description{
This internal function scans the CIGAR string of a single aligned read to
determine if it contains a specific insertion or deletion at a precise genomic location.
}
\keyword{internal}
