% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample_data.R
\name{summarise_cohort}
\alias{summarise_cohort}
\title{Summarise cohort data
This function is intended to summarise cohort data.}
\usage{
summarise_cohort(samples, samplesheet)
}
\arguments{
\item{samples}{A list of samples matrices loaded from the samplesheet.}

\item{samplesheet}{A tibble containing the samplesheet data.}
}
\description{
Summarise cohort data
This function is intended to summarise cohort data.
}
