% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_plot.R
\name{target_plot}
\alias{target_plot}
\title{Plot Targets by Sample}
\usage{
target_plot(target, palette = "Set1")
}
\description{
This function plots target data for
a specified target and source sample.
the color palette to use for the plot.
}
