% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-accepted-ref-alt.R
\name{trim_left_prefix}
\alias{trim_left_prefix}
\title{Trim Common Left Prefix from Two Sequences}
\usage{
trim_left_prefix(a, b)
}
\arguments{
\item{a}{A character string representing the first sequence.}

\item{b}{A character string representing the second sequence.}
}
\value{
A list with two elements:
The modified first sequence with minimal shared prefix.}
The modified second sequence with minimal shared prefix.}
}
\description{
This function trims any common left prefix between two sequences while ensuring
that at least one shared base remains. If a common prefix is found, the function
keeps exactly one shared base and removes the rest of the overlapping prefix.
}
