% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-normalize-bcftools_norm.R
\name{apply_bcftools_norm}
\alias{apply_bcftools_norm}
\title{Normalize VCF file using bcftools norm
This function normalizes a VCF file by calling bcftools norm on a temp VCF.
The normalized output is then read into a dataframe, extracting the columns:
chr, pos, ref, and alt.}
\usage{
apply_bcftools_norm(chr, pos, ref, alt, fasta, tmp_folder)
}
\arguments{
\item{chr}{Character vector representing the chromosome of interest.}

\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{fasta}{A reference genome in FASTA format.}

\item{tmp_folder}{Character vector for the folder temporary path.}
}
\value{
A dataframe containing normalized variant information with bcftools.
}
\description{
Normalize VCF file using bcftools norm
This function normalizes a VCF file by calling bcftools norm on a temp VCF.
The normalized output is then read into a dataframe, extracting the columns:
chr, pos, ref, and alt.
}
\keyword{internal}
