% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_bcftools_norm.R
\name{apply_bcftools_norm}
\alias{apply_bcftools_norm}
\title{Normalize a single variant using bcftools norm}
\usage{
apply_bcftools_norm(chr, pos, ref, alt, fasta, tmp_folder, verbose)
}
\arguments{
\item{chr}{A string representing the chromosome.}

\item{pos}{An integer representing the position.}

\item{ref}{A string representing the reference allele.}

\item{alt}{A string representing the alternative allele.}

\item{fasta}{Path to the FASTA file for the reference sequence used for generating the BAM file.}

\item{tmp_folder}{Character vector for the temporary folder path.}

\item{verbose}{Boolean. If set to TRUE, print all the warnings and the prints.}
}
\value{
A dataframe containing normalized variant information with bcftools.
}
\description{
This function normalizes a single variant by leveraging the
external 'bcftools norm' command. It writes the variant to a temporary VCF
file, executes 'bcftools norm' for left-alignment and parsimonious
representation, and then reads the normalized result back into a data frame.
}
\keyword{internal}
