% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_fragments_info_table.R
\name{build_fragments_info_table}
\alias{build_fragments_info_table}
\title{Create a table with statistics about fragments covering a position of interest (e.g the position of a mutation).}
\usage{
build_fragments_info_table(df_sam, chr, pos, n_cores = 1)
}
\arguments{
\item{df_sam}{dataframe containing reads from SAM format}

\item{chr}{chromosome for position of interest}

\item{pos}{chromosome for position of interest}

\item{n_cores}{the number of cores available}
}
\value{
a dataframe
}
\description{
Create a table with statistics about fragments covering a position of interest (e.g the position of a mutation).
}
