% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-check_parameters.R
\name{check_parameters}
\alias{check_parameters}
\title{Check input files for fRagmentomics function}
\usage{
check_parameters(
  mut,
  bam,
  fasta,
  sample_id,
  neg_offset_mate_search,
  pos_offset_mate_search,
  one_based,
  flag_bam_list,
  report_tlen,
  report_softclip,
  report_5p_3p_bases_fragment,
  cigar_free_indel_match,
  tmp_folder,
  output_file,
  n_cores
)
}
\arguments{
\item{mut}{Path to a .vcf or .tsv file or string representation chr:pos:ref:alt of a mutation.}

\item{bam}{Path to a BAM file.}

\item{fasta}{Path to the FASTA file for the reference sequence used for generating the BAM file.}

\item{sample_id}{Sample identifier.}

\item{neg_offset_mate_search}{Integer. Use in read_bam.
Represents the number of nucleotides to extend upstream (negative direction)
from the position of interest when querying the BAM file with Rsamtools.
his extension ensures that paired reads are retrieved, even if only one mate
overlaps the queried position.}

\item{pos_offset_mate_search}{Integer. Use in read_bam.}

\item{one_based}{Boolean. TRUE if fasta is in one based. False if in 0 based.}

\item{flag_bam_list}{A named list of logicals for filtering reads based on their SAM flag
NA = Filter is ignored, TRUE = The read MUST have this flag, FALSE = The read MUST NOT have this flag.}

\item{report_tlen}{Boolean. Whether to include the TLEN (template length)
information in the output.}

\item{report_softclip}{Boolean. Whether to include the number of soft-clipped
bases at the fragment extremities in the output.}

\item{report_5p_3p_bases_fragment}{Integer. Whether to include N fragment
extremity bases in the output.}

\item{cigar_free_indel_match}{Boolean. If activated, the information from the CIGAR is disregarded when determining the
mutation status of a read for indel. Instead the mutation status is determined by comparing the sequence of the read
to the sequence of the wild-type reference and the mutated reference. Activating this option may lead to discordant
genotyping of reads compared to the information provided by the CIGAR for indels. On the other hand, when
activated, it may rescue mutated genotypes for indel that would be missed in cases where the representation of the
indel in the CIGAR does not match the norm of bcftools of the mutation being analyzed.}

\item{tmp_folder}{Character vector for the folder temporary path.}

\item{output_file}{Character vector for the output file path.}

\item{n_cores}{Number of cores for parallel computation.}
}
\value{
None. The function stops execution if error.
}
\description{
Check input files for fRagmentomics function
}
\keyword{internal}
