% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_read_to_ref_wt_and_mut.R
\name{compare_read_to_ref_wt_and_mut}
\alias{compare_read_to_ref_wt_and_mut}
\title{Compare a read sequence against reference and alternate alleles}
\usage{
compare_read_to_ref_wt_and_mut(
  read_seq,
  ref_seq_wt,
  ref_seq_mut,
  compare_len_wt,
  compare_len_mut
)
}
\arguments{
\item{read_seq}{Base sequence of the read}

\item{ref_seq_wt}{Base sequence of the wild-type ref}

\item{ref_seq_mut}{Base sequence of the mutated ref}

\item{compare_len_wt}{Number of bases to be compared in read vs wild-type ref comparison}

\item{compare_len_mut}{Number of bases to be compared in read vs mutated ref comparison}
}
\value{
a character describing the mutation status of the read
}
\description{
A core comparison utility that classifies a read sequence by checking its compatibility with the
expected wild-type (WT) and mutant (MUT) sequences.
}
\keyword{internal}
