% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fragment_features.R
\name{extract_fragment_features}
\alias{extract_fragment_features}
\title{Process a single DNA fragment from paired-end reads}
\usage{
extract_fragment_features(
  df_sam,
  fragment_name,
  sample_id,
  chr,
  pos,
  ref,
  alt,
  report_bam_info,
  report_softclip,
  report_5p_3p_bases_fragment,
  remove_softclip,
  fasta_fafile = NULL,
  fasta_seq = NULL,
  input_mutation_info
)
}
\arguments{
\item{df_sam}{A dataframe containing sequencing reads.}

\item{fragment_name}{Name of the fragment (paired-end reads).}

\item{sample_id}{Sample identifier.}

\item{chr}{Character vector representing the chromosome of interest.}

\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{report_bam_info}{Boolean. Whether to include the bam information.}

\item{report_softclip}{Boolean. Whether to include the number of soft-clipped
bases at the fragment extremities in the output.}

\item{report_5p_3p_bases_fragment}{Integer. Whether to include N fragment
extremity bases in the output.}

\item{remove_softclip}{Boolean. If set to TRUE, trim soft-clipped bases from the 5' end of Read 5p and from the 3' end of Read 3p.}

\item{fasta_fafile}{An open connection to an object of class FaFile}

\item{fasta_seq}{A list with the fasta sequence between two positions.}

\item{input_mutation_info}{Character vecotr representing the input mutation.}
}
\value{
A dataframe with the processed fragment information.
}
\description{
This is a high-level worker function that orchestrates the
complete analysis of a single DNA fragment (represented by a pair of reads)
at a specific variant locus. It performs quality control, identifies reads,
extracts features, determines mutation status, and returns all results in a
structured format.
}
\details{
The function executes the following pipeline for each fragment:
\enumerate{
\item It subsets the reads from 'df_sam' that match the given 'fragment_name'.
\item It performs initial quality control checks (e.g., proper pairing,
chromosome consistency) via 'process_fragment_reads_qc'.
\item It identifies the 5' (forward strand) and 3' (reverse strand) reads
based on their SAM FLAGs.
\item If 'remove_softclip' is 'TRUE', it trims soft-clipped bases from the
sequences, qualities, and CIGAR strings.
\item It calls 'get_base_basq_mstat_from_read' to determine the allele and
mutation status for each individual read at the variant position.
\item It calculates the precise fragment size using 'get_fragment_size'.
\item It consolidates the two read statuses into a final fragment status
(e.g., 'MUT', 'DISCORDANT') using 'get_mutation_status_of_fragment'.
\item It assembles and returns a single-row data frame containing all
extracted information.
}
}
\keyword{internal}
