% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRagmentomics.R
\name{fRagmentomics}
\alias{fRagmentomics}
\title{A user-friendly R package, that enables the characterization
of each cfDNA fragment overlapping one or multiple mutations of interest,
starting from a sequencing file containing aligned reads (BAM file).}
\description{
fRagmentomics supports multiple mutation input formats (e.g.,
VCF, TSV, or string “chr:pos:ref:alt” representation), accommodates one-based
and zero-based genomic conventions, handles mutation representation ambiguities,
and accepts any reference file and species in FASTA format. For each cfDNA
fragment, fRagmentomics outputs its size, its 3’ and 5’ sequences, and its
mutational status.
}
\author{
Killian Maudet
}
