% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fRagmentomics.R
\name{fRagmentomics}
\alias{fRagmentomics}
\title{fRagmentomics}
\usage{
fRagmentomics(
  bam,
  mut,
  fasta,
  sample = NA,
  neg_offset_mate_search = -1000,
  pos_offset_mate_search = 1000,
  one_based = TRUE,
  flag_keep = 3,
  flag_remove = 2304,
  report_5p_bases = 0,
  report_3p_bases = 0,
  report_tlen = FALSE,
  report_softclip = FALSE,
  n_cores = 1
)
}
\arguments{
\item{n_cores}{Number of cores for parallel computation.}

\item{df_sam}{A dataframe containing sequencing reads.}

\item{sample_id}{Sample identifier.}

\item{chr}{Chromosome of interest.}

\item{pos}{Genomic position of interest.}

\item{ref}{Reference base.}

\item{alt}{Alternative base (for SNV) or sequence (for insertion).}

\item{mutation_type}{Type of mutation: "mutation", "deletion", or "insertion".}
}
\value{
A dataframe containing extracted fragment-level information.
}
\description{
Processes fragment-level sequencing data in parallel and extracts relevant statistics.
}
