% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_base_basq_mstat_from_read.R
\name{get_base_basq_mstat_from_read}
\alias{get_base_basq_mstat_from_read}
\title{Get read sequence, read base qualities, and read mutational status}
\usage{
get_base_basq_mstat_from_read(
  chr,
  pos,
  ref,
  alt,
  read_stats,
  fasta_fafile = NULL,
  fasta_seq = NULL,
  cigar_free_indel_match = FALSE
)
}
\arguments{
\item{chr}{Character vector representing the chromosome of interest.}

\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{read_stats}{A list of read-level statistics.}

\item{fasta_fafile}{An open connection to an object of class FaFile}

\item{fasta_seq}{A list with the fasta sequence between two positions.}

\item{cigar_free_indel_match}{Boolean. If activated, the information from the CIGAR is disregarded when determining the
mutation status of a read for indel. Instead the mutation status is determined by comparing the sequence of the read
to the sequence of the wild-type reference and the mutated reference. Activating this option may lead to discordant
genotyping of reads compared to the information provided by the CIGAR for indels. On the other hand, when
activated, it may rescue mutated genotypes for indel that would be missed in cases where the representation of the
indel in the CIGAR does not match the norm of bcftools of the mutation being analyzed.}
}
\value{
A list containing "base", "basq", "mstat".
}
\description{
Get read sequence, read base qualities, and read mutational status
}
\keyword{internal}
