% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fragment_bases_5p_3p_softclip.R
\name{get_fragment_bases_5p_3p_softclip}
\alias{get_fragment_bases_5p_3p_softclip}
\title{Extract Soft-Clipped Base Counts from CIGAR Strings}
\usage{
get_fragment_bases_5p_3p_softclip(cigar_5p, cigar_3p)
}
\arguments{
\item{cigar_5p}{Character string. The CIGAR string of the 5' read.}

\item{cigar_3p}{Character string. The CIGAR string of the 3' read.}
}
\value{
A 'list' containing two integer:
\itemize{
\item 'nb_softclip_5p': The number of soft-clipped bases at the start of
the 5' read.
\item 'nb_softclip_3p': The number of soft-clipped bases at the end of
the 3' read.
}
}
\description{
This function extracts the number of soft-clipped bases at the 5' ends of 5' reads and 3' ends of 3' reads.
If no soft clipping is detected, the count is set to 0.
}
\keyword{internal}
