% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_fragment_inner_distance.R
\name{get_fragment_inner_distance}
\alias{get_fragment_inner_distance}
\title{Compute fragment inner distance
Calculates the inner distance between two paired-end sequencing reads.}
\usage{
get_fragment_inner_distance(
  r_pos1,
  r_cigar1,
  read_length1,
  r_pos2,
  r_cigar2,
  read_length2
)
}
\arguments{
\item{r_pos1}{numeric value representing the read 1 mapping position}

\item{r_cigar1}{character vector representing the read 1 CIGAR}

\item{read_length1}{numeric value representing the length of read 1}

\item{r_pos2}{numeric value representing the read 2 mapping position}

\item{r_cigar2}{character vector representing the read 2 CIGAR}

\item{read_length2}{numeric value representing the length of read 2}
}
\value{
a named list with names \code{a} and \code{b}.
}
\description{
Compute fragment inner distance
Calculates the inner distance between two paired-end sequencing reads.
}
\keyword{internal}
