% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-get_base_basq_mstat_from_read.R
\name{get_index_aligning_with_pos}
\alias{get_index_aligning_with_pos}
\title{Get the index of the element in the sequence aligning with the position of interest.}
\usage{
get_index_aligning_with_pos(pos, read_stats)
}
\arguments{
\item{pos}{Numeric value representing the Genomic position of interest.}

\item{read_stats}{A list of read-level statistics.}
}
\value{
An integer representing the index of the nucleotide in sequence aligning with the position of interest. If
the read does not cover the position of interest, this integer is -1. If the read contains a deletion or skipping at
the position of interest, this integer is -2.
}
\description{
Extracts base, quality, and indel informations based on the mutation type.
}
\keyword{internal}
