% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector-insertion.R
\name{get_insertion}
\alias{get_insertion}
\title{Get the informations about the presence of a insertion in the read.}
\usage{
get_insertion(pos, alt, r_pos, r_cigar, r_query, r_qual)
}
\arguments{
\item{pos}{numeric value representing the position of interest}

\item{alt}{character vector representing the insertion sequence}

\item{r_pos}{numeric value representing the read mapping position}

\item{r_cigar}{character vector representing the read CIGAR}

\item{r_query}{character vector representing the read base sequence}

\item{r_qual}{character vector representing the read sequencing qualities}
}
\value{
a named list with names \code{base},\code{qual}.
}
\description{
Get the informations about the presence of a insertion in the read.
}
\keyword{internal}
