% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutation_status_of_read.R
\name{get_number_of_common_first_char}
\alias{get_number_of_common_first_char}
\title{Count common leading characters between two strings}
\usage{
get_number_of_common_first_char(str_a, str_b)
}
\arguments{
\item{str_a}{a string}

\item{str_b}{a string}
}
\value{
An integer
}
\description{
Calculates the length of the common prefix for two given character strings.
}
\keyword{internal}
