% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_plot.R
\name{housekeeping_plot}
\alias{housekeeping_plot}
\title{Plot Housekeeping Genes}
\usage{
housekeeping_plot(housekeeping, random, relative = TRUE)
}
\arguments{
\item{housekeeping}{A tibble containing the housekeeping gene peak data.}

\item{random}{A tibble containing the random datasets peak data.}

\item{relative}{A boolean indicating whether to plot relative coverage}
}
\description{
This function generates a ggplot plot for housekeeping genes.
}
\examples{
\dontrun{
example_samplesheet <- system.file("extdata",
                                   "samplesheet.csv",
                                   package = "fragmentomics")
samplesheet <- parse_samplesheet(example_samplesheet)
experiment <- load_experiment(samplesheet, "results")
housekeeping <- experiment |> dplyr::filter(target_label == "GeneHancer_housekeeping")
housekeeping_data <- load_peaks(housekeeping, "results")
random <- experiment |> dplyr::filter(source_label == "random_dataset")
random_data <- load_peaks(random, "results")
housekeeping_plot(housekeeping_data, random_data)
}
}
