% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_plot.R
\name{housekeeping_plot}
\alias{housekeeping_plot}
\title{Plot Housekeeping Genes}
\usage{
housekeeping_plot(housekeeping_data, random_data, relative = FALSE)
}
\arguments{
\item{housekeeping_data}{A tibble containing
the housekeeping_data gene peak data.}

\item{random_data}{A tibble containing the random_data datasets peak data.}

\item{relative}{A boolean indicating whether to plot relative coverage}
}
\description{
This function generates a ggplot plot for housekeeping genes.
}
\examples{
\dontrun{
example_samplesheet <- system.file("extdata",
                                   "samplesheet.csv",
                                   package = "fragmentomics")
samplesheet <- parse_samplesheet(example_samplesheet)
experiment <- load_experiment(samplesheet, "results")
housekeeping <- experiment |>
  dplyr::filter(target_label == "GeneHancer_housekeeping",
                sampleid == "SAMPLE_1_BL")
housekeeping_data <- load_peaks(housekeeping, "results")
random <- experiment |> dplyr::filter(source_label == "random_dataset",
                                      sampleid == "SAMPLE_1_BL")
random_data <- load_peaks(random, "results")
housekeeping_plot(housekeeping_data, random_data)
}
}
