% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_plot.R
\name{housekeeping_plot_relative}
\alias{housekeeping_plot_relative}
\title{Plot Relative Housekeeping Genes}
\usage{
housekeeping_plot_relative(housekeeping_data, random_data)
}
\arguments{
\item{housekeeping_data}{A tibble containing the housekeeping_data gene peak data.}

\item{random_data}{A tibble containing the random_data datasets peak data.}
}
\description{
Plot Relative Housekeeping Genes
}
