% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experiment.R
\name{load_experiment}
\alias{load_experiment}
\title{Load sample data from a samplesheet
This function is intended to load sample data
based on a provided samplesheet.}
\usage{
load_experiment(
  samplesheet,
  rootpath,
  subdir = "fragmentomics/processed/matrix",
  parallelize = TRUE,
  number_of_daemons = parallel::detectCores()
)
}
\arguments{
\item{samplesheet}{A character string specifying
the path to the samplesheet file.}

\item{rootpath}{A character string specifying the
root path where the sample data is located.}

\item{subdir}{A character string specifying the
subdirectory within the root path where
the sample data is located, defaults
to "fragmentomics/processed/matrix".}

\item{parallelize}{A logical value indicating whether to parallelize the
loading of sample data, defaults to TRUE.#' loading of sample data, defaults to TRUE.}

\item{number_of_daemons}{An integer specifying the number of daemons to use}
}
\value{
A list containing the loaded sample data.
}
\description{
Load sample data from a samplesheet
This function is intended to load sample data
based on a provided samplesheet.
}
