% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sample.R
\name{load_sample}
\alias{load_sample}
\title{Load data files from a sample directory}
\usage{
load_sample(caseid, sampleid, timepoint, encoded_timepoint, rootpath, subdir)
}
\arguments{
\item{caseid}{A character string specifying the case ID.}

\item{sampleid}{A character string specifying the sample ID.}

\item{timepoint}{A factor specifying the timepoint.}

\item{encoded_timepoint}{An integer specifying the encoded timepoint.}

\item{rootpath}{A character string specifying the root path where the sample
data is located.}

\item{subdir}{A character string specifying the subdirectory
within the root path where
the sample data is located, defaults to "fragmentomics/processed".}
}
\description{
This function is intended to load data files from a
specified sample directory.
}
