% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-normalize_mut.R
\name{normalize_mut}
\alias{normalize_mut}
\title{Check input files for fRagmentomics function}
\usage{
normalize_mut(df_mut, fasta, fasta_fafile, one_based)
}
\arguments{
\item{df_mut}{a dataframe with mutations}

\item{fasta}{Path to the FASTA file for the reference sequence used for generating the BAM file.}

\item{fasta_fafile}{An open connection to an object of class FaFile}

\item{one_based}{Boolean. TRUE if fasta is in one based. False if in 0 based.}
}
\value{
a dataframe with normalized mutations representations containing columns CHROM, POS, REF, ALT
}
\description{
Check input files for fRagmentomics function
}
\keyword{internal}
