% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_to_vcf_rep.R
\name{normalize_to_vcf_rep}
\alias{normalize_to_vcf_rep}
\title{Normalize a variant to VCF representation}
\usage{
normalize_to_vcf_rep(chr, pos, ref, alt, fasta_fafile, one_based, verbose)
}
\arguments{
\item{chr}{A string representing the chromosome.}

\item{pos}{An integer representing the position.}

\item{ref}{A string representing the reference allele.}

\item{alt}{A string representing the alternative allele.}

\item{fasta_fafile}{An open connection to an object of class FaFile}

\item{one_based}{Boolean. TRUE if fasta is in one based. False if in 0 based.}

\item{verbose}{Boolean. If set to TRUE, print all the warnings and the prints.}
}
\value{
A list with chr, pos, ref and alt
}
\description{
This function converts a variant from a user-provided format into the
standard VCF representation, which is crucial for handling indels. It also
harmonizes chromosome names and validates the reference allele against a
FASTA file.
}
\details{
The normalization process involves several key steps:
\enumerate{
\item \strong{Coordinate Adjustment}: Converts the input 'pos' from 0-based to 1-based if 'one_based' is 'FALSE'.
\item \strong{Chromosome Naming}: Standardizes the chromosome name (e.g., '1' vs 'chr1') to match the provided FASTA
reference using 'harmonize_chr_to_fasta'.
\item \strong{Indel Padding}: For insertions and deletions, it prepends an "anchor" base from the reference genome to both
'ref' and 'alt' alleles to create a valid VCF-style representation. The 'pos' is adjusted accordingly for
deletions. SNVs and MNVs are not modified.
\item \strong{Validation}: After normalization, it confirms that the final 'ref' allele matches the sequence in the FASTA file at the new coordinates.
}
}
\keyword{internal}
