% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-fragment.R
\name{process_fragment}
\alias{process_fragment}
\title{Process a single sequencing fragment
Extracts and processes relevant data from a sequencing fragment.}
\usage{
process_fragment(
  df_sam,
  fragment_name,
  sample_id,
  chr,
  pos,
  ref,
  alt,
  pos_after_indel_repetition,
  mutation_type,
  report_tlen,
  report_softclip,
  report_5p_3p_bases_fragment
)
}
\arguments{
\item{df_sam}{A dataframe containing sequencing reads.}

\item{fragment_name}{Name of the fragment (paired-end reads).}

\item{sample_id}{Sample identifier.}

\item{chr}{Character vector representing the chromosome of interest.}

\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{pos_after_indel_repetition}{Integer. The position after the indel
repetion.}

\item{mutation_type}{In "SNV", "deletion", or "insertion".}

\item{report_tlen}{Boolean. Whether to include the TLEN (template length)
information in the output.}

\item{report_softclip}{Boolean. Whether to include the number of soft-clipped
bases at the fragment extremities in the output.}

\item{report_5p_3p_bases_fragment}{Integer. Whether to include N fragment
extremity bases in the output.}
}
\value{
A dataframe with the processed fragment information.
}
\description{
Process a single sequencing fragment
Extracts and processes relevant data from a sequencing fragment.
}
