% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-read_bam.R
\name{read_bam}
\alias{read_bam}
\title{Preprocess BAM file to extract reads around a given position
This function extracts reads from a BAM file using Rsamtools, filtering them
based on flags and selecting reads of interest around a genomic position.}
\usage{
read_bam(
  bam,
  chr,
  pos,
  neg_offset_mate_search,
  pos_offset_mate_search,
  flag_bam_list
)
}
\arguments{
\item{bam}{Path to a BAM file.}

\item{chr}{Character. Chromosome of interest.}

\item{pos}{Integer. Genomic position of interest.}

\item{neg_offset_mate_search}{Integer. Use in read_bam.
Represents the number of nucleotides to extend upstream (negative direction)
from the position of interest when querying the BAM file with Rsamtools.
his extension ensures that paired reads are retrieved, even if only one mate
overlaps the queried position.}

\item{pos_offset_mate_search}{Integer. Use in read_bam.}

\item{flag_bam_list}{A named list of logicals for filtering reads based on their SAM flag
NA = Filter is ignored, TRUE = The read MUST have this flag, FALSE = The read MUST NOT have this flag.}
}
\value{
A dataframe containing the filtered SAM entries.
}
\description{
Preprocess BAM file to extract reads around a given position
This function extracts reads from a BAM file using Rsamtools, filtering them
based on flags and selecting reads of interest around a genomic position.
}
\keyword{internal}
