% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-read_mut.R
\name{read_mut}
\alias{read_mut}
\title{Read mutations input into a dataframe. Multiallelic mutations
are split into multiple rows.}
\usage{
read_mut(mut)
}
\arguments{
\item{mut}{Path to a .vcf or .tsv file or string representation chr:pos:ref:alt of a mutation.}
}
\value{
a df with only biallelic mutations containing columns CHROM, POS, REF, ALT
}
\description{
Read mutations input into a dataframe. Multiallelic mutations
are split into multiple rows.
}
\keyword{internal}
