% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepro-read-mut.R
\name{read_mut}
\alias{read_mut}
\title{Put the informations about mutations into a dataframe}
\usage{
read_mut(mut)
}
\value{
a df without multiallelic ALT and without header
with chr pos ref alt columns
}
\description{
Put the informations about mutations into a dataframe
}
\keyword{internal}
