% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{scale_x_bins}
\alias{scale_x_bins}
\title{scale_x_bins}
\usage{
scale_x_bins(bin_vector, central_bin, bin_size)
}
\value{
A list with breaks and labels for the x scale.
}
\description{
Function to build the x scale
}
