% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-search_for_indel_in_cigar.R
\name{search_for_indel_in_cigar}
\alias{search_for_indel_in_cigar}
\title{Get the informations about the presence of a insertion in the read.}
\usage{
search_for_indel_in_cigar(pos, ref, alt, read_stats, type)
}
\arguments{
\item{pos}{Numeric value representing the Genomic position of interest.}

\item{ref}{Character vector representing reference base(s).}

\item{alt}{Character vector representing alternative base(s).}

\item{read_stats}{A list of read-level statistics.}

\item{type, }{choose between "INS" or "DEL"}
}
\value{
a boolean indicating TRUE if the insertion was found in the CIGAR a the position of interest, FALSE otherwise
}
\description{
Get the informations about the presence of a insertion in the read.
}
\keyword{internal}
