% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_plot.R
\name{target_plot}
\alias{target_plot}
\title{Plot Targets by Sample}
\usage{
target_plot(target, palette = "Set1")
}
\arguments{
\item{target}{A tibble containing target data}

\item{palette}{A character string specifying the color palette to use.
Defaults to "Set1" from RColorBrewer.}
}
\value{
A ggplot object visualizing the target data.
}
\description{
This function plots target data for
a specified target and source sample.
}
\examples{
\dontrun{
example_samplesheet <- system.file("extdata",
                                   "samplesheet.csv",
                                   package = "fragmentomics")
samplesheet <- parse_samplesheet(example_samplesheet)
experiment <- load_experiment(samplesheet, "results")
ctcf <- experiment |> dplyr::filter(target_label == "CTCF")
ctcf_peaks <- load_peaks(, "results")
target_plot(ctcf_peaks)
}
}
