% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_plot.R
\name{target_plot}
\alias{target_plot}
\title{Plot Targets by Sample}
\usage{
target_plot(experiment, target, source, palette = "Set1")
}
\arguments{
\item{experiment}{An experiment object containing target data.}

\item{target}{A character string specifying the target to plot.}

\item{source}{A character string specifying the source sample to plot.}

\item{palette}{A character string specifying the color palette to use for the plot.}
}
\description{
This function plots target data for a specified target and source sample.
}
